Move datastore creation to own module
Move to a core:preferences module so the call to create can be reused.
