+++ core/preferences/build.gradle.kts
@@ -5,7 +5,6 @@ plugins {
     alias(libs.plugins.kotlin.multiplatform)
     alias(libs.plugins.kotlin.serialization)
     alias(libs.plugins.ksp)
-    alias(libs.plugins.ktorfit)
 }
 
 kotlin {
@@ -25,7 +24,7 @@ kotlin {
         iosSimulatorArm64()
     ).forEach {
         it.binaries.framework {
-            baseName = "auth"
+            baseName = "preferences"
             isStatic = true
         }
     }
@@ -34,39 +33,19 @@ kotlin {
         androidMain.dependencies {
         }
         commonMain.dependencies {
-            implementation(projects.core.network)
-            implementation(projects.core.preferences)
             implementation(libs.androidx.datastore.preferences)
             implementation(libs.koin.core)
             implementation(libs.kotlinx.coroutines.core)
-            implementation(libs.kotlinx.serialization.json)
-            implementation(libs.ktor.client.content.negotiation)
-            implementation(libs.ktor.client.serialization)
-            implementation(libs.ktorfit.converters.response)
-            implementation(libs.ktorfit.lib)
-            implementation(libs.ktor.serialization.kotlinx.json)
         }
         commonTest.dependencies {
-            implementation(libs.kotlin.test)
-            implementation(libs.kotlinx.coroutines.test)
         }
         iosMain.dependencies {
         }
     }
 }
 
-dependencies {
-    with(libs.ktorfit.ksp) {
-        add("kspAndroid", this)
-        add("kspCommonMainMetadata", this)
-        add("kspIosArm64", this)
-        add("kspIosSimulatorArm64", this)
-        add("kspIosX64", this)
-    }
-}
-
 android {
-    namespace = "com.chesire.nekomp.library.datasource.auth"
+    namespace = "com.chesire.nekomp.core.network"
     compileSdk = 35
     defaultConfig {
         minSdk = 27
