+++ core/preferences/src/androidMain/kotlin/com/chesire/nekomp/core/preferences/CreateDataStore.android.kt
@@ -1,0 +1,8 @@
+package com.chesire.nekomp.core.preferences
+
+import android.content.Context
+import org.koin.java.KoinJavaComponent.get
+
+internal actual fun producePath(filename: String): String {
+    return get<Context>(Context::class.java).filesDir.resolve(filename).absolutePath
+}
