+++ core/preferences/src/commonMain/kotlin/com/chesire/nekomp/core/preferences/CreateDataStore.kt
@@ -1,0 +1,14 @@
+package com.chesire.nekomp.core.preferences
+
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.PreferenceDataStoreFactory
+import androidx.datastore.preferences.core.Preferences
+import okio.Path.Companion.toPath
+
+fun createDataStore(filename: String): DataStore<Preferences> {
+    return PreferenceDataStoreFactory.createWithPath(
+        produceFile = { producePath(filename).toPath() }
+    )
+}
+
+internal expect fun producePath(filename: String): String
