--- library/datasource/auth/src/androidMain/kotlin/com/chesire/nekomp/library/datasource/auth/local/CreateDataStore.android.kt
@@ -1,8 +1,0 @@
-package com.chesire.nekomp.library.datasource.auth.local
-
-import android.content.Context
-import org.koin.java.KoinJavaComponent.get
-
-internal actual fun producePath(): String {
-    return get<Context>(Context::class.java).filesDir.resolve(AUTH_DATASTORE_NAME).absolutePath
-}
