+++ library/datasource/auth/src/commonMain/kotlin/com/chesire/nekomp/library/datasource/auth/LibraryAuthModule.kt
@@ -1,8 +1,8 @@
 package com.chesire.nekomp.library.datasource.auth
 
 import com.chesire.nekomp.core.network.ResultConverterFactory
+import com.chesire.nekomp.core.preferences.createDataStore
 import com.chesire.nekomp.library.datasource.auth.local.AuthStorage
-import com.chesire.nekomp.library.datasource.auth.local.createDataStore
 import com.chesire.nekomp.library.datasource.auth.remote.AuthApi
 import com.chesire.nekomp.library.datasource.auth.remote.createAuthApi
 import de.jensklingenberg.ktorfit.ktorfitBuilder
@@ -12,6 +12,8 @@ import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
+
+private const val AUTH_DATASTORE_NAME = "auth.preferences_pb"
 
 val libraryAuthModule = module {
     factory<AuthApi> {
@@ -34,6 +36,6 @@ val libraryAuthModule = module {
     }
     singleOf(::AuthRepository)
     single<AuthStorage> {
-        AuthStorage(createDataStore())
+        AuthStorage(createDataStore(AUTH_DATASTORE_NAME))
     }
 }
